Move log_submission_info into DriverLogger
